+++ tests/Parts/Channel/ChannelTest.php
@@ -16,7 +16,7 @@ use Discord\Discord;
 use Discord\Helpers\Collection;
 use Discord\Parts\Channel\Channel;
 use Discord\Parts\Channel\Message;
-use Discord\Parts\Guild\Invite;
+use Discord\Parts\Channel\Invite;
 
 /**
  * @covers \Discord\Parts\Channel\Channel
